Grand Addition
Added finally the true logic of the attacks, also added the EnemyData and AttackData as scriptable objects per enemy, so we can have a better management of the information of those.

Finished the startup of the battles, now when doing contact with a enemy, it will inicialize a sequence on a coroutine to mantain the cycles of the enemies attacks.

Player data isnt made for now, so only the enemy logic has been done (which was tbh the easier part) now we have to advance to player logic like his stats, his attacks, parrys, dodges, etc.

Right now i only made 1 enemy with 1 attack to test, he could have more attacks with different stats every attack.

Adjusted the battle zone a little from the previous one (i realized i made it too big)

Also there is a bug with the player when tping, for some reason it stays on top of the platform, most possibly smth to do with how i made the movement on the first version, ill remake it later.